refactor(bindinginfo): move symbols to their own package
BindingInfo is going to be consumed from a different package than
servicebindingrequest, this change is required to avoid circular dependencies
and to keep internals private at the same time.
