+++ pkg/controller/servicebindingrequest/bindinginfo/bindinginfo.go
@@ -1,8 +1,13 @@
-package servicebindingrequest
+package bindinginfo
 
 import (
-	"fmt"
+	"errors"
 	"strings"
+)
+
+var (
+	// ServiceBindingOperatorAnnotationPrefix is the prefix of Service Binding Operator related annotations.
+	ServiceBindingOperatorAnnotationPrefix = "servicebindingoperator.redhat.io/"
 )
 
 // BindingInfo represents the pieces of a binding as parsed from an annotation.
@@ -14,10 +19,20 @@ type BindingInfo struct {
 	Path string
 	// Descriptor is the field reference to another manifest
 	Descriptor string
+	// Value is the ...
+	Value string
 }
+
+var InvalidAnnotationPrefixErr = errors.New("invalid annotation prefix")
+var InvalidAnnotationNameErr = errors.New("invalid annotation name")
 
 // NewBindingInfo parses the encoded in the annotation name, returning its parts.
 func NewBindingInfo(name string, value string) (*BindingInfo, error) {
+	// do not process unknown annotations
+	if !strings.HasPrefix(name, ServiceBindingOperatorAnnotationPrefix) {
+		return nil, InvalidAnnotationPrefixErr
+	}
+
 	cleanName := strings.TrimPrefix(name, ServiceBindingOperatorAnnotationPrefix)
 	parts := strings.SplitN(cleanName, "-", 2)
 
@@ -27,6 +42,7 @@ func NewBindingInfo(name string, value string) (*BindingInfo, error) {
 			FieldPath:  parts[0],
 			Path:       parts[0],
 			Descriptor: strings.Join([]string{value, parts[0]}, ":"),
+			Value:      value,
 		}, nil
 	}
 
@@ -36,8 +52,9 @@ func NewBindingInfo(name string, value string) (*BindingInfo, error) {
 			FieldPath:  parts[0],
 			Path:       parts[1],
 			Descriptor: strings.Join([]string{value, parts[1]}, ":"),
+			Value:      value,
 		}, nil
 	}
 
-	return nil, fmt.Errorf("should have two parts")
+	return nil, InvalidAnnotationNameErr
 }
