+++ pkg/controller/servicebindingrequest/bindinginfo/bindinginfo_test.go
@@ -1,15 +1,16 @@
-package servicebindingrequest
+package bindinginfo
 
 import (
-	"github.com/stretchr/testify/require"
 	"testing"
+
+	"github.com/stretchr/testify/require"
 )
 
 // TestNewBindingInfo exercises annotation binding information parsing.
 func TestNewBindingInfo(t *testing.T) {
 	type args struct {
-		s string
-		d string
+		name  string
+		value string
 	}
 	tests := []struct {
 		name    string
@@ -18,21 +19,23 @@ func TestNewBindingInfo(t *testing.T) {
 		want    *BindingInfo
 	}{
 		{
-			args: args{s: "status.configMapRef-password", d: "binding"},
+			args: args{name: "status.configMapRef-password", value: "binding"},
 			want: &BindingInfo{
 				FieldPath:  "status.configMapRef",
 				Descriptor: "binding:password",
 				Path:       "password",
+				Value:      "binding",
 			},
 			name:    "{fieldPath}-{path} annotation",
 			wantErr: false,
 		},
 		{
-			args: args{s: "status.connectionString", d: "binding"},
+			args: args{name: "status.connectionString", value: "binding"},
 			want: &BindingInfo{
 				Descriptor: "binding:status.connectionString",
 				FieldPath:  "status.connectionString",
 				Path:       "status.connectionString",
+				Value:      "binding",
 			},
 			name:    "{path} annotation",
 			wantErr: false,
@@ -40,7 +43,7 @@ func TestNewBindingInfo(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			b, err := NewBindingInfo(tt.args.s, tt.args.d)
+			b, err := NewBindingInfo(tt.args.name, tt.args.value)
 			if err != nil && !tt.wantErr {
 				t.Errorf("NewBindingInfo() error = %v, wantErr %v", err, tt.wantErr)
 				return
